Create layout for wild ttt
