+++ app/src/main/AndroidManifest.xml
@@ -19,7 +19,7 @@
 
         <activity android:name=".AboutActivity"/>
         <activity android:name=".RandomActivity"/>
-        <activity android:name=".WildActivity"/>
+        <activity android:name=".oldWildActivity"/>
     </application>
 
 </manifest>
